requirements.txt
transformers>=4.44.0
accelerate>=0.33.0
sentence-transformers>=3.0.0
langdetect>=1.0.9
pyyaml>=6.0.1
tqdm>=4.66.0
uvicorn>=0.30.0
fastapi>=0.115.0
python-dotenv>=1.0.1
rich>=13.7.0
faiss-cpu>=1.7.4
deep_translator

